+++ NEWS.md
@@ -32,6 +32,7 @@ Red Panda C++ Version 2.0
   - change: Don't generate localized filename when new header/add file in the project view
   - fix: Restore project's original compiler set if user choose 'No' in the confirm project compiler set change dialog.
   - fix: Encoding info in the status bar not correctly updated when save a new file
+  - enhancement: auto sort TODO items 
 
 Red Panda C++ Version 1.5